+++ spec/lib/polish_geeks/dev_tools/shell_spec.rb
@@ -1,0 +1,13 @@
+require 'spec_helper'
+
+RSpec.describe PolishGeeks::DevTools::Shell do
+  subject { described_class.new }
+
+  describe '#execute' do
+    context 'when we run a command' do
+      it 'executes the command' do
+        expect(subject.execute('ls')).to include('Gemfile')
+      end
+    end
+  end
+end